+++ tests/testthat/test-crew_controller_group.R
@@ -13,7 +13,6 @@ crew_test("crew_controller_group() active bindings for covr", {
   x <- crew_controller_local()
   y <- crew_controller_group(x = x)
   expect_true(inherits(y$relay, "crew_class_relay"))
-  expect_true(inherits(y$throttle, "crew_class_throttle"))
 })
 
 crew_test("crew_controller_group()", {